% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Create 'bubble' plot}
\usage{
bubble_plot(plot_df, x = "x", y = "y", size = "size",
  fill = "fill", x_labels = NULL, y_labels = NULL, ...)
}
\arguments{
\item{plot_df}{data.frame  data to plot}

\item{x}{character   Name of the variable to plot on the x-axis}

\item{y}{character   Name of the variable to plot on the y-axis}

\item{size}{character   Name of the variable to use for the size of the
points}

\item{fill}{character   Name of the variable to use for the fill colour}

\item{x_labels}{character   Optional labels to use instead of levels of
x-axis variable}

\item{y_labels}{character   Optional labels to use instead of levels of
y-axis variable}
}
\value{
plot - ggplot2 object
}
\description{
\code{bubble_plot} produces a ggplot bubble plot from a data.frame
}
\details{
This function takes a data.frame and creates a bubble plot.
An x-y scatterplot where an attribute is represented by the size of the
points and optionally an another attribute is mapped to the colour.
}
\examples{

set.seed(20962)
test_data_cat <- data.frame(x = factor(rep(LETTERS[1:10], 10)),
  y = factor(rep(1:10, each = 10), levels = 10:1),
  size = sample(20:100, 100, replace = TRUE),
  fill = runif(100))

cat_bubble_plot <- bubble_plot(test_data_cat)

test_data_num <- test_data_cat
test_data_num$x <- sample(1:50, 20)
test_data_num$y <- sample(1:50, 20)

num_bubble_plot <- bubble_plot(test_data_num)

}
