% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendro_plot}
\alias{dendro_plot}
\title{Create dendrogram from a clustering object}
\usage{
dendro_plot(hclust_obj, ...)
}
\arguments{
\item{hclust_obj}{hclust  tree to plot}
}
\value{
plot - ggplot2 object
}
\description{
\code{dendro_plot} produces a ggplot dendrogram from an hclust object
}
\details{
This function takes an hclust object and use ggdendro to produce
a ggplot dendrogram.
}
\examples{


}
