% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gg_legend.R
\name{get_gg_legend}
\alias{get_gg_legend}
\title{Get the legend of a plot as a grob}
\usage{
get_gg_legend(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{ggplot object - the plot to get the legend from}
}
\value{
grob
}
\description{
\code{get_gg_legend} returns the legend of a ggplot object as a grob
}
\examples{

legend <- get_gg_legend(ggplot_obj)

}
