% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_heatmap.R
\name{matrix_heatmap}
\alias{matrix_heatmap}
\title{Create heatmap plot for a matrix}
\usage{
matrix_heatmap(
  data_matrix,
  x_title = "Sample",
  y_title = "Gene",
  fill_title = "Value",
  fill_palette = "plasma",
  xaxis_labels = TRUE,
  yaxis_labels = TRUE,
  colour = NULL,
  size = NULL,
  na.translate = TRUE,
  ...
)
}
\arguments{
\item{x_title}{character   Name for the x-axis}

\item{y_title}{character   Name for the y-axis}

\item{fill_title}{character   Name for the legend}

\item{fill_palette}{character   Name of the fill palette to use.
Must be one of the viridis or color brewer palettes}

\item{xaxis_labels}{logical/character   default=TRUE, FALSE means labels are not printed
if the value is a character vector of the same length as the levels of the x variable
these values are used instead}

\item{yaxis_labels}{logical/character   default=TRUE, FALSE means labels are not printed
if the value is a character vector of the same length as the levels of the y variable
these values are used instead}

\item{colour}{character  fixed value for the border colour of the heatmap tiles}

\item{size}{integer  fixed value for the line width of the heatmap tiles border}

\item{na.translate}{logical  whether to include NA values in the legend for categorical fill variables}

\item{...}{Other arguments passed on to \code{\link{theme_heatmap()}}}

\item{plot_matrix}{matrix      data to plot}
}
\value{
plot - ggplot2 object
}
\description{
\code{matrix_heatmap} produces a ggplot heatmap from a matrix
}
\details{
This function takes a matrix and creates a heatmap plot.
The default values for the x and y axis titles are Sample and Gene
respectively and the fill legend is labelled Value.
By default x and y axis labels are printed.
}
\examples{
set.seed(1638)
test_matrix <- matrix(
  sample(1:16),
  nrow = 4
)
matrix_heatmap(test_matrix)

# specify x, y and fill titles and fill palette
matrix_heatmap(test_matrix, x_title = "Cols", y_title = "Rows",
               fill_title = "Log2FC", fill_palette = "inferno")

# supply different labels for x and y axes
matrix_heatmap(test_matrix, x_title = "Sample", y = "Sample_2",
               fill_title = "Log2FC",
               fill_palette = "magma", xaxis_labels = letters[1:4],
               yaxis_labels = LETTERS[1:4])

# remove x and y labels
matrix_heatmap(test_matrix, xaxis_labels = FALSE, yaxis_labels = FALSE)

}
