% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_with_fill_and_shape.R
\name{scatterplot_with_fill_and_shape}
\alias{scatterplot_with_fill_and_shape}
\title{Plot 2 columns against each other with Fill and Shapes}
\usage{
scatterplot_with_fill_and_shape(
  plot_data,
  x_var,
  y_var,
  fill_var,
  fill_palette,
  shape_var = "None",
  shape_palette,
  point_labels = NULL,
  label_seed = NA,
  point_size = 4,
  ...
)
}
\arguments{
\item{plot_data}{data.frame - plot data}

\item{x_var}{character - name of the column to plot on the x axis}

\item{y_var}{character - name of the column to plot on the y axis}

\item{fill_var}{character - name of the column to use as the fill aesthetic}

\item{fill_palette}{character - a named character vectors of colours for the fill aesthetic}

\item{shape_var}{character - name of the column to use as the shape aesthetic}

\item{shape_palette}{character - a named character vectors of shapes for the shape aesthetic}

\item{point_labels}{character - labels for the points. Either a column name
from the plot_data or a character vector. Default: NULL - no labels}

\item{point_size}{numeric - size of points}

\item{labels_seed}{integer - seed for geom_text_repel}
}
\value{
ggplot2 object
}
\description{
\code{scatterplot_with_fill_and_shape} Takes a data.frame and creates an x-y scatterplot
}
\details{
The plot data data.frame should contain the columns for the x, y and fill variables
       and optionally ones for the shape variable and text labels.
       The text label column should be named sample_name.
}
\examples{
scatterplot_with_fill_and_shape( plot_data, 'PC1', 'PC2', 'Gene', fill_palette,
                             'Genotype', shape_palette, sample_names )

}
