% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{theme_bubble}
\alias{theme_bubble}
\title{A ggplot theme for bubble plots}
\usage{
theme_bubble(base_size = 11, base_family = "", categorical = TRUE)
}
\arguments{
\item{base_size}{numeric    base size of text in points (default: 11)}

\item{base_family}{character  base font family (default: "")}

\item{categorical}{logical    are either of the x and y variable categorical (default: TRUE)}
}
\value{
ggplot theme object
}
\description{
\code{theme_bubble} produces a ggplot bubble plot from a data.frame
}
\details{
This is a ggplot theme for bubble plots.
For categorical axes, it is theme_void with x and y axis labels
It also has grey dotted grid lines and the legend is place at the top

If both x and y variables are continuous, theme_minimal is returned

\code{bubble_plot} uses this theme internally
}
\examples{

set.seed(20962)
test_data_cat <- data.frame(x = factor(rep(LETTERS[1:10], 10)),
  y = factor(rep(1:10, each = 10), levels = 10:1),
  size = sample(20:100, 100, replace = TRUE),
  fill = runif(100))

cat_bubble_plot <- bubble_plot(test_data_cat)
cat_bubble_plot + theme_bubble() + theme(axis.text.x = element_text(angle = 0, face = 'italic'))

}
