% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_heatmap.R
\name{theme_heatmap}
\alias{theme_heatmap}
\title{A ggplot theme for heatmap plots}
\usage{
theme_heatmap(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ...
)
}
\arguments{
\item{base_size}{numeric    base size of text in points (default: 11)}

\item{base_family}{character  base font family (default: "")}

\item{categorical}{logical    are either of the x and y variable categorical (default: TRUE)}
}
\value{
ggplot theme object
}
\description{
\code{theme_heatmap} produces a ggplot heatmap plot from a data.frame
}
\details{
This is a ggplot theme for heatmap plots.
For categorical axes, it is theme_void with x and y axis labels
It also has grey dotted grid lines and the legend is place at the top

If both x and y variables are continuous, theme_minimal is returned
}
\examples{

set.seed(20962)
test_data_cat <- data.frame(x = factor(rep(LETTERS[1:10], 10)),
  y = factor(rep(1:10, each = 10), levels = 10:1),
  size = sample(20:100, 100, replace = TRUE),
  fill = runif(100))

cat_heatmap_plot <- heatmap_plot(test_data_cat)
cat_heatmap_plot + theme_heatmap() + theme(axis.text.x = element_text(angle = 0, face = 'italic'))

}
