% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendro_plot}
\alias{dendro_plot}
\title{Create dendrogram from a clustering object}
\usage{
dendro_plot(hclust_obj, categorical_scale = TRUE)
}
\arguments{
\item{hclust_obj}{hclust  tree to plot}

\item{categorical_scale}{logical whether the x axis should be categorical or not
This affects the space around the axes and whether the tree plot will line up with
other plots with categorical axes}
}
\value{
plot - ggplot2 object
}
\description{
\code{dendro_plot} produces a ggplot dendrogram from an hclust object
}
\details{
This function takes an hclust object and uses ggdendro to produce
a ggplot dendrogram.
}
\examples{

dendro_plot(cluster_res$clustering)

}
